requirements.txt
streamlit==1.24.0
pandas==2.0.3
Pillow==9.5.0
pdfkit==1.0.0
python-docx==0.8.11